Rename variables, fix zod schemas

+++ backend/src/validationSchemas/city.schema.ts
@@ -7,11 +7,11 @@ export const AddCitySchema = z.object({
 export const GetCitiesSchema = z.object({
     limit: z.preprocess(
         (a) => parseInt(z.string().parse(a), 10),
-        z.number().positive()
+        z.number().int().positive()
     ).optional(),
     page: z.preprocess(
         (a) => parseInt(z.string().parse(a), 10),
-        z.number().positive()
+        z.number().int().positive()
     ).optional()
 });
 export const GetCitySchema = z.object({
